Modified the database helper and dbManager classes.
column changes made to the Dataset table: [Where - = Removed, + = Added]
 - DATASET_ID (Local ID)
 + DS_COL_DESCRIPTION

column changes made to the Image table:
 - IMAGE_THUMBNAIL_PATH
 - IMAGE_COL_DATASET_ID
 + IMAGE_COL_CACHE_PATH
 + IMAGE_COL_DATASET_NAME

Added default values to columns in the case they are not provided when creating entries.

The primary keys for the image table have been changed to be the image name and the dataset name. This is because the same image could potentially exist in different datasets, however, we won't allow for duplicates in the same dataset.

Added dataset functions to:
- get datasetPK
- insert local datasets (in offline mode)
- update local datasets
- update local dataset after sync
- delete local dataset
- insert synced dataset (in online mode)
- update dataset
- delete synced dataset
- get/set dataset sync status
- get unsynced datasets
- insert synced image
- insert unsynced image
- insert/get image cache path
- reclassify image
- delete image
- get cachedImageList
- get unsynced images