Modified the NetworkTask constructor to take the dataset name as the input parameter instead of the datasetPK as a dataset PK is not allocated until a dataset has been synced.
The reclassify and upload task have been to modified to deal with online and offline mode. In offline mode, dataset changes will be stored to the local database. In online mode, dataset changes will be handled using the restAPI and then information in the local database will be updated accordingly.