Added a network connectivity monitor to observe offline/online mode changes. If we are currently viewing a dataset which has been synced in the background, the changes made to the dataset will be updated in the local database.
This will also handle the case in which, if a user has entered a dataset where no images are loaded whilst in offline mode, and then the phone connects to a network, the process images function will be called automatically. Hence, there is no reason to exit and re-enter the dataset to load it's contents.

Added a function to cache compressed images as they are processed.