The image is now being loaded into the holder using a filepath. First we check if the image has been cached, if so, then retrieve cached image. Else, retrieve full image.
If an image is clicked while action mode if off, open it up in an image viewer. If the user swipes the image viewer, transition to the previous image or next image depending on swipe direction. If a user swipes in a vertical manner, then dismiss the image viewer and return to the image grid recyclerview.