Added a function to download dataset contents (images) for offline mode use. Users will be presented the option to download full images or compressed images in the confirmation dialog.
Added a network connectivity monitor to monitor changes in connectivity. When in online mode, enable the features to copy datasets and download datasets.

Modified the add, delete and edit dataset function to work for offline mode and online mode usage. During offline mode, changes will be stored in the local database. Whereas, in online mode, an API request will be made and then the local database will be updated accordingly.